requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
mistralai>=0.0.7
requests>=2.28.0
geopy>=2.2.0